+++ src/id_translation/fetching/_pandas_fetcher.py
@@ -1,6 +1,5 @@
-import os
 from pathlib import Path
-from typing import Any, Callable, Dict, List, Mapping, Union
+from typing import Any, Callable, Dict, Iterable, List, Mapping, Union
 
 import pandas as pd
 from rics._internal_support.types import PathLikeType
@@ -11,8 +10,8 @@ from ..types import IdType
 from ._abstract_fetcher import AbstractFetcher
 from .types import FetchInstruction
 
-PandasReadFunction = Callable[[PathLikeType], pd.DataFrame]
-FormatFn = Callable[[PathLikeType], str]
+PandasReadFunction = Callable[[str], pd.DataFrame]
+FormatFn = Callable[[str], str]
 
 
 class PandasFetcher(AbstractFetcher[str, IdType]):
@@ -22,24 +21,30 @@ class PandasFetcher(AbstractFetcher[str, IdType]):
     a Pandas function such as :func:`pandas.read_csv` or :func:`pandas.read_parquet`, but any function that accepts a
     string `source` as the  first argument and returns a data frame can be used.
 
+    .. hint::
+
+       When using **remote file systems**, :attr:`~.AbstractFetcher.sources` are resolved using
+       `AbstractFileSystem.glob()`_. If resolution fails, consider overriding the :meth:`find_sources`-method.
+
     Args:
         read_function: A Pandas `read`-function. If a string is given, the function is resolved using
             :func:`rics.misc.get_by_full_name`. Unqualified names are assumed to belong to the ``pandas`` namespace.
-        read_path_format: A formatting string or a callable to apply to a source before passing them to `read_function`.
-            Must contain a `source` as its only placeholder. Example: ``data/{source}.pkl``. Leave as-is if ``None``.
-            Valid URL schemes include http, ftp, s3, gs, and file.
+        read_path_format: A string on the form ``protocol://path/to/sources/{}.ext``, or a callable to apply
+            to a source before passing them to `read_function`.
         read_function_kwargs: Additional keyword arguments for `read_function`.
         online: Setting ``online=False`` typically indicates that files are hosted at a location where there are access
             limitations, e.g. through data transfer fees.
 
     See Also:
         The official `Pandas IO documentation <https://pandas.pydata.org/pandas-docs/stable/user_guide/io.html>`_
+
+    .. _AbstractFileSystem.glob(): https://filesystem-spec.readthedocs.io/en/latest/api.html?highlight=glob#fsspec.spec.AbstractFileSystem.glob
     """
 
     def __init__(
         self,
-        read_function: Union[PandasReadFunction, str] = "read_pickle",
-        read_path_format: Union[str, FormatFn] = "data/{}.pkl",
+        read_function: Union[PandasReadFunction, str] = "read_csv",
+        read_path_format: Union[str, FormatFn] = "data/{}.csv",
         read_function_kwargs: Mapping[str, Any] = None,
         online: bool = False,
         **kwargs: Any,
@@ -54,7 +59,7 @@ class PandasFetcher(AbstractFetcher[str, IdType]):
         self._online = online
         self._kwargs = read_function_kwargs or {}
 
-        self._source_paths: Dict[str, Path] = {}
+        self._source_paths: Dict[str, str] = {}
 
     def read(self, source_path: PathLikeType) -> pd.DataFrame:
         """Read a ``DataFrame`` from a source path.
@@ -65,35 +70,65 @@ class PandasFetcher(AbstractFetcher[str, IdType]):
         Returns:
             A deserialized ``DataFrame``.
         """
-        return self._read(source_path, **self._kwargs)
+        return self._read(source_path, **self._kwargs).convert_dtypes()
 
-    def find_sources(self) -> Dict[str, Path]:
-        """Search for source paths to pass to `read_function` using `read_path_format`.
+    def format_source(self, source: str) -> str:
+        """Get the path for `source`."""
+        return self._format_source(source)
+
+    def find_sources(self) -> Dict[str, str]:
+        """Resolve sources and their associated paths.
+
+        Sources are resolved in three steps:
+
+        1. Create glob pattern by calling :meth:`format_source` with ``source='*'``.
+        2. Glob files using `AbstractFileSystem.glob()`_ (requires ``fsspec``) or :meth:`Path.glob() <pathlib.Path.glob>`.
+        3. Strip the directory  and file suffix from the globbed paths to create source names.
+
+        .. _AbstractFileSystem.glob(): https://filesystem-spec.readthedocs.io/en/latest/api.html?highlight=glob#fsspec.spec.AbstractFileSystem.glob
 
         Returns:
             A dict ``{source, path}``.
-
-        Raises:
-            IOError: If files cannot be read.
         """
-        abs_file = Path(self._format_source("")).absolute()
-        directory = abs_file.parent
-        file_pattern = abs_file.name
+        pattern = self.format_source("*")
 
-        if not directory.is_dir():  # pragma: no cover
-            problem = "is not a directory" if directory.exists() else "does not exist"
-            raise IOError(f"Bad path format: {directory} {problem}.")
-
-        source_paths = {}
-        # Path.glob does not work with absolute directories.
-        for file in map(Path, os.listdir(directory)):
-            if file.name.endswith(str(file_pattern)):  # pragma: no cover
-                source_paths[file.name.replace(file_pattern, "")] = directory.joinpath(file)
+        try:
+            source_paths = self._find_sources_fsspec(pattern)
+        except ModuleNotFoundError as e:
+            self.logger.debug(f"Falling back to 'pathlib.Path': {e!r}")
+            source_paths = self._find_sources_pathlib(pattern)
 
         if not source_paths:  # pragma: no cover
-            pattern = Path(self._format_source("*")).absolute()
-            raise IOError(f"Bad path pattern: '{pattern}' did not match any files.")
+            self.logger.warning(f"Bad path pattern: '{pattern}' did not match any files.")
 
+        self.logger.debug(f"Source paths resolved: {source_paths}")
+
+        return source_paths
+
+    def _find_sources_fsspec(self, pattern: str) -> Dict[str, str]:
+        from fsspec.core import url_to_fs  # type: ignore
+
+        fs, _ = url_to_fs(pattern, **self._kwargs.get("storage_options", {}))
+
+        source_paths = self._make_source_paths(fs.glob(pattern))
+
+        protocol, separator, _ = pattern.partition("://")
+        if separator:
+            source_paths = {source: protocol + separator + path for source, path in source_paths.items()}
+
+        return source_paths
+
+    def _find_sources_pathlib(self, pattern: str) -> Dict[str, str]:
+        path = Path(pattern)
+        iterator = path.parent.glob(path.name)
+        return self._make_source_paths(iterator)
+
+    @staticmethod
+    def _make_source_paths(iterator: Iterable[Path]) -> Dict[str, str]:
+        source_paths = {}
+        for path in map(Path, iterator):
+            source = path.with_suffix("").name
+            source_paths[source] = str(path)
         return source_paths
 
     def _initialize_sources(self, task_id: int) -> Dict[str, List[str]]:
@@ -111,5 +146,5 @@ class PandasFetcher(AbstractFetcher[str, IdType]):
         return self._online
 
     def __repr__(self) -> str:
-        read_path_format = self._format_source("{source}")
+        read_path_format = self.format_source("{}")
         return f"{tname(self)}(read_function={tname(self._read)}, {read_path_format=})"